chore: rename function
