--- packages/picasso/src/CarouselGradient/utils/calculateGradientWidth/index.ts
@@ -1,1 +1,0 @@
-export { default } from './calculateGradientWidth'
