+++ packages/picasso/src/CarouselGradient/utils/gradientWidth/gradientWidth.ts
