+++ packages/picasso/src/CarouselGradient/utils/gradientWidth/index.ts
@@ -1,0 +1,1 @@
+export { default } from './gradientWidth'
